refactor: move Network + SafeHttp under App\Services\Network
Two structural changes:

1. SafeHttp grew well past a Helper — it's a stateful, SSRF-hardened
   HTTP client wrapper. Move it (and Network, its tight collaborator)
   into App\Services\Network. Tests follow under Tests\Unit\Services\Network.

2. Drop app(SafeHttp::class) service-locator call inside EpisodePlayable.
   The value class now takes SafeHttp as a parameter on
   getForEpisode(\$episode, \$safeHttp); the two callers
   (PodcastStreamerAdapter, DownloadService) inject SafeHttp via
   constructor and pass it through.

EpisodePlayableTest initializes one SafeHttp instance in setUp() and
reuses it across tests instead of resolving in each body.
