+++ tests/Integration/Values/EpisodePlayableTest.php
@@ -4,6 +4,7 @@ namespace Tests\Integration\Values;
 
 use App\Exceptions\UnsafeUrlException;
 use App\Models\Song;
+use App\Services\Network\SafeHttp;
 use App\Values\Podcast\EpisodePlayable;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Http;
@@ -12,6 +13,15 @@ use Tests\TestCase;
 
 class EpisodePlayableTest extends TestCase
 {
+    private SafeHttp $safeHttp;
+
+    public function setUp(): void
+    {
+        parent::setUp();
+
+        $this->safeHttp = app(SafeHttp::class);
+    }
+
     #[Test]
     public function createAndRetrieved(): void
     {
@@ -24,14 +34,14 @@ class EpisodePlayableTest extends TestCase
                 'path' => 'https://example.com/episode.mp3',
             ]);
 
-        $playable = EpisodePlayable::getForEpisode($episode);
+        $playable = EpisodePlayable::getForEpisode($episode, $this->safeHttp);
 
         Http::assertSentCount(1);
         self::assertSame('acbd18db4cc2f85cedef654fccc4a4d8', $playable->checksum);
 
         self::assertTrue(Cache::has("episode-playable.{$episode->id}"));
 
-        $retrieved = EpisodePlayable::getForEpisode($episode);
+        $retrieved = EpisodePlayable::getForEpisode($episode, $this->safeHttp);
 
         // No extra HTTP request should be made.
         Http::assertSentCount(1);
@@ -56,7 +66,7 @@ class EpisodePlayableTest extends TestCase
         self::expectException(UnsafeUrlException::class);
 
         try {
-            EpisodePlayable::getForEpisode($episode);
+            EpisodePlayable::getForEpisode($episode, $this->safeHttp);
         } finally {
             Http::assertNothingSent();
         }
@@ -77,7 +87,7 @@ class EpisodePlayableTest extends TestCase
         self::expectException(UnsafeUrlException::class);
 
         try {
-            EpisodePlayable::getForEpisode($episode);
+            EpisodePlayable::getForEpisode($episode, $this->safeHttp);
         } finally {
             // The redirect target must never be requested.
             Http::assertNotSent(static fn ($request) => str_contains($request->url(), '169.254.169.254'));
